# specs/001-trail-campsite-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Hiking Trail MVP API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
...
  /trails:
    get:
      summary: List trails with search and pagination
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200': { description: Trail list response }
  /trails/{trailId}:
    get:
      summary: Get trail detail
      parameters:
        - in: path
          name: trailId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Trail detail response }
        '404': { description: Trail not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT